fix(core): fix race-condition during reload
*Motivation*

We found a race-condition affecting Beamium during reload when a config
file is changed. You will find a log report on gist on
https://gist.github.com/PierreZ/76a309ffada4ece9fa520b3c17fe1a12#file-beamium-concurrency-issue-log-L23-L26

The main issue was that when receiving multiples watches events, the
control loop was not synchronized with the state of the cmd::main
function, and the shared variable used to sync was modified within a
thread::spawn during a former loop iteration. This was the worflow with
two events queued:

  1. first iteration is fine, firing a thread::spawn
  2. second iteration is almost immediate after 1, setting signal to false
  3. we are now waiting for thread to join
  4. BUT, between 2 and 3, Thread::spawn made progress, set it to true

As false is required to allow the thread to join, we are observing
a deadlock on the control loop.

We added two features to avoid this:

  * a shared AtomicBool which will provide sync between control loop and
cmd::main. With this, we will be able to wait for cmd::main to be
actually started before starting a new iteration. Failure is still
handled in the same way.

  * a drain of the watch's iterator. Instead of taking events one by
one and reload for each events, we are trying the drain the iterator.
This is useful when mounting /etc/beamium as a Kubernetes's ConfigMap,
as any modifications to the ConfigMap will generate at least 6 events.